requirements.txt
mysql-connector-python==8.0.33
pillow==9.5.0
tkcalendar==1.6.1
python-dotenv==1.0.0
cryptography==41.0.1
bcrypt==4.0.1
pandas==2.1.0
openpyxl==3.1.2
reportlab==4.0.4
python-dateutil==2.8.2
